+++ src/Eawv.Service/Models/NotificationServiceModels/EmailSenderModel.cs
@@ -1,9 +1,0 @@
-﻿// (c) Copyright by Abraxas Informatik AG
-// For license information see LICENSE file
-
-namespace Eawv.Service.Models.NotificationServiceModels;
-
-public class EmailSenderModel
-{
-    public string DisplayName { get; set; }
-}