+++ src/Eawv.Service/Services/INotificationService.cs
@@ -9,8 +9,7 @@ using Eawv.Service.Models.TemplateServiceModels;
 namespace Eawv.Service.Services;
 
 /// <summary>
-/// Service for sending notifications using the SEAS Notification Service.
-/// Used version / endpoint and genaral configuration see <see cref="NotificationServiceConfiguration"/>.
+/// Service for sending notification emails via the configured <see cref="Voting.Lib.UserNotifications.IUserNotificationSender"/>.
 /// </summary>
 public interface INotificationService
 {
@@ -18,11 +17,11 @@ public interface INotificationService
     /// Renders the template and sends it as an email via the notification service.
     /// The filename of the template is used as subject.
     /// </summary>
-    /// <param name="recipientLoginIds">The Secure Connect login IDs of the targeted recipients which should receive an email.</param>
+    /// <param name="recipientEmails">The email addresses of the targeted recipients which should receive an email.</param>
     /// <param name="type">The template type for the email.</param>
     /// <param name="bag">The template data model.</param>
     /// <returns>Task for async handling.</returns>
-    Task SendEmailAsync(List<string> recipientLoginIds, TemplateType type, TemplateBag bag);
+    Task SendEmailAsync(List<string> recipientEmails, TemplateType type, TemplateBag bag);
 
     /// <summary>
     /// Sends an email using the provided request model.